+++ modulo5/projeto-labook/requests.rest
@@ -5,8 +5,8 @@ POST http://localhost:3003/signup
 Content-Type: application/json
 
 {
-   "name": "matheus" ,
-   "email": "matheus@gmail.com" ,
+   "name": "jose" ,
+   "email": "jose@gmail.com" ,
    "password": "123456"
 }
 
@@ -17,7 +17,7 @@ POST http://localhost:3003/login
 Content-Type: application/json
 
 {
-   "email": "matheus@gmail.com" ,
+   "email": "jose@gmail.com" ,
    "password": "123456"
 }
 
@@ -30,15 +30,15 @@ Authorization: {{token}}
 
 {
    "photo": "link da foto" ,
-   "description": "Carnaval no Rio",
-   "date": "20/03/1999",
+   "description": "Carnaval em SP",
+   "date": "20/03/2019",
    "role": "evento"
 }
 
 ###
 @token = {{ login.response.body.token }}
 
-@id = 0bae13e0-1e4e-4dbd-ad59-059985cd038b
+@id = 03ecb6d4-2c4e-49c8-9acd-d6b0cb0bc896
 # Pegar Post por ID 
 GET  http://localhost:3003/post/{{id}}
 Authorization: {{token}}
@@ -53,7 +53,7 @@ Content-Type: application/json
 Authorization: {{token}}
 
 {
-   "id": "ef2efe9a-860e-46e8-8fbb-2f89b2cf0201"
+   "id": "1dbdce1b-44ea-49db-aeb5-c233e90816e9"
 }
 
 ###
@@ -71,14 +71,14 @@ Authorization: {{token}}
 @token = {{ login.response.body.token }}
 
 # Pegar post que usuário segue
-GET  http://localhost:3003/user/posts
+GET  http://localhost:3003/posts
 Authorization: {{token}}
 
 ###
 @token = {{ login.response.body.token }}
 
 # Pegar Post por Type
-GET  http://localhost:3003/user/normal
+GET  http://localhost:3003/user/evento
 Authorization: {{token}}
