+++ modulo5/projeto-labook/src/controller/UserController.ts
@@ -1,5 +1,6 @@
 import { Request, Response } from "express";
 import UserBusiness from "../business/UserBusiness";
+import { FriendInputDTO } from "../types/friendInputDTO";
 import { LoginInputDTO } from "../types/loginInputDTO";
 import { SignupInputDTO } from "../types/singupInputDTO";
 
@@ -47,4 +48,40 @@ export default class UserController {
       res.status(500).send("Internal server error");
     }
   }
+
+  followFriend = async (req: Request, res: Response) => {
+    try {
+      const token = req.headers.authorization!;
+      const id = req.body;
+
+      const followId: FriendInputDTO = id;
+
+      const result = await this.userBusiness.follow(token, followId);
+
+      res.status(201).send(result);
+    } catch (error) {
+      if (error instanceof Error) {
+        return res.status(400).send(error.message);
+      }
+      res.status(500).send("Internal server error");
+    }
+  };
+
+  unfollowFriend = async (req: Request, res: Response) => {
+    try {
+      const token = req.headers.authorization!;
+      const {id} = req.body;
+
+      const unFollowId: FriendInputDTO = {id};
+
+      const result = await this.userBusiness.deleteFallow(token, unFollowId);
+
+      res.status(201).send(result);
+    } catch (error) {
+      if (error instanceof Error) {
+        return res.status(400).send(error.message);
+      }
+      res.status(500).send("Internal server error");
+    }
+  };
 }
