+++ modulo5/projeto-labook/src/data/PostData.ts
@@ -35,6 +35,26 @@ export default class PostData extends BaseDataBase {
     }
   }
 
+  public postFollow = async (id_follow: string) => {
+    try {
+      const result = await BaseDataBase.connection(this.TABLE_NAME)
+      .select("labook_users.name", "labook_users.email", "labook_posts.photo", "labook_posts.description", "labook_posts.created_at")
+      .join("labook_users","labook_users.id", "labook_friends.id_followed")
+      .join("labook_posts","labook_posts.author_id", "labook_friends.id_followed")
+      .where({id_follow})
+      .orderBy('created_at', 'desc')
+      .into("labook_friends")
+      return result
+
+    } catch (error) {
+      if (error instanceof Error) {
+        throw new Error(error.message);
+      } else {
+        throw new Error("Erro do banco !");
+      }
+    }
+  };
+
   public postType =async (type:string) => {
     try {
       const result = await BaseDataBase.connection(this.TABLE_NAME)
