+++ modulo5/projeto-labook/src/index.ts
@@ -1,14 +1,8 @@
-import FeedBusiness from "./business/FeedBusiness";
-import FriendBusiness from "./business/FriendBusiness";
 import PostBusiness from "./business/PostBusiness";
 import UserBusiness from "./business/UserBusiness";
 import { app } from "./controller/app";
-import FeedController from "./controller/FeedController";
-import FriendController from "./controller/FriendController";
 import PostController from "./controller/PostController";
 import UserController from "./controller/UserController";
-import FeedData from "./data/FeedData";
-import FriendData from "./data/FriendData";
 import PostData from "./data/PostData";
 import UserData from "./data/UserData";
 
@@ -18,12 +12,6 @@ const userController = new UserController(userBusiness)
 const postBusiness = new PostBusiness(new PostData())
 const postController = new PostController(postBusiness)
 
-const friendBusiness = new FriendBusiness(new FriendData())
-const friendController = new FriendController(friendBusiness)
-
-const feedBusiness = new FeedBusiness(new FeedData())
-const feedController = new FeedController(feedBusiness)
-
 
 
 app.post("/signup", userController.singUp )
@@ -31,13 +19,13 @@ app.post("/login", userController.login )
 app.post("/post", postController.createPost)
 
 app.get("/post/:id", postController.getPostById)
-app.get("/user/:type", postController.getPosByType)
-app.get("/user/posts", feedController.postUserFollow)
+app.get("/user/:type", postController.getPostByType)
+app.get("/posts", postController.postUserFollow)
 
 
-app.put("/user/follow", friendController.followFriend)
+app.put("/user/follow", userController.followFriend)
 
-app.delete("/user/unfollow", friendController.unfollowFriend)
+app.delete("/user/unfollow", userController.unfollowFriend)
